requirements.txt
slack-sdk==3.23.0
aiohttp==3.9.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
redis==5.0.0
asyncio==3.4.3
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
msgraph-core==1.0.0
azure-identity==1.15.0
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.9.0
flake8==6.1.0